+++ modules/core/src/scenegraph/model.js
@@ -2,28 +2,15 @@ import GL from '@luma.gl/constants';
 import BaseModel from './base-model';
 import Attribute from '../core/attribute';
 import {getDrawMode} from '../core/geometry';
-import {Buffer, Query, Program, TransformFeedback} from '../webgl';
-import {MODULAR_SHADERS, assembleShaders} from '@luma.gl/shadertools';
-import {logModel} from '../debug/seer-integration';
+import {Buffer, Query, TransformFeedback} from '../webgl';
+
 // import {removeModel} from '../debug/seer-integration';
 import {log, isObjectEmpty} from '../utils';
 import assert from '../utils/assert';
 
-import {getDebugTableForUniforms} from '../webgl-debug/debug-uniforms';
-import {getDebugTableForVertexArray} from '../webgl-debug/debug-vertex-array';
-import {getDebugTableForProgramConfiguration} from '../webgl-debug/debug-program-configuration';
-
 const ERR_MODEL_PARAMS = 'Model needs drawMode and vertexCount';
-
 const LOG_DRAW_PRIORITY = 2;
-const LOG_DRAW_TIMEOUT = 10000;
 
-// These old picking uniforms should be avoided and we should use picking module
-// and set uniforms using Model class 'updateModuleSettings()'
-// TODO - move to shader modules
-const DEPRECATED_PICKING_UNIFORMS = ['renderPickingBuffer', 'pickingEnabled'];
-
-// Model abstract O3D Class
 export default class Model extends BaseModel {
   constructor(gl, props = {}) {
     super(gl, props);
@@ -209,113 +196,18 @@ export default class Model extends BaseModel {
 
   // PRIVATE METHODS
 
-  _createProgram({
-    vs = null,
-    fs = null,
-    // 1: Modular shaders
-    modules = null,
-    defines = {},
-    inject = {},
-    shaderCache = null,
-    // TransformFeedback
-    varyings = null,
-    bufferMode = GL.SEPARATE_ATTRIBS,
-    program = null
-  }) {
-    this.getModuleUniforms = x => {};
-
-    const id = this.id;
-
-    if (!program) {
-      // Assign default shaders if none are provided
-      vs = vs || MODULAR_SHADERS.vs;
-      fs = fs || MODULAR_SHADERS.fs;
-
-      const assembleResult = assembleShaders(this.gl, {vs, fs, modules, inject, defines, log});
-      ({vs, fs} = assembleResult);
-
-      if (shaderCache) {
-        program = shaderCache.getProgram(this.gl, {id, vs, fs});
-      } else {
-        program = new Program(this.gl, {id, vs, fs, varyings, bufferMode});
-      }
-
-      this.getModuleUniforms = assembleResult.getUniforms || (x => {});
+  // Updates (evaluates) all function valued uniforms based on a new set of animationProps
+  // experimental
+  _setAnimationProps(animationProps) {
+    if (this.animated) {
+      assert(animationProps, 'Model.draw(): animated uniforms but no animationProps');
+      const animatedUniforms = this._evaluateAnimateUniforms(animationProps);
+      this.program.setUniforms(animatedUniforms, {}, () => {
+        // if something changed
+        this._checkForDeprecatedUniforms(animatedUniforms);
+        this.setNeedsRedraw();
+      });
     }
-
-    assert(program instanceof Program, 'Model needs a program');
-    return program;
-  }
-  /* eslint-enable complexity */
-
-  // Uniforms
-
-  _checkForDeprecatedUniforms(uniforms) {
-    // deprecated picking uniforms
-    DEPRECATED_PICKING_UNIFORMS.forEach(uniform => {
-      if (uniform in uniforms) {
-        log.deprecated(
-          uniform,
-          'use picking shader module and Model class updateModuleSettings()'
-        )();
-      }
-    });
-  }
-
-  // Refreshes animated uniforms, attempting to get animated props from animationLoop if registered
-  _refreshAnimationProps(animationProps) {
-    // Try to read animationProps
-    animationProps = animationProps || (this.animationLoop && this.animationLoop.animationProps);
-    if (animationProps) {
-      this._setAnimationProps(animationProps);
-    }
-  }
-
-  // Calculate new values for any function uniforms based on supplied animationProps
-  _evaluateAnimateUniforms(animationProps) {
-    if (!this.animated) {
-      return {};
-    }
-    const animatedUniforms = {};
-    for (const uniformName in this.animatedUniforms) {
-      const valueFunction = this.animatedUniforms[uniformName];
-      animatedUniforms[uniformName] = valueFunction(animationProps);
-    }
-    return animatedUniforms;
-  }
-
-  // Extracts a list of function valued uniforms, so we can update them before each draw call
-  // Also removes such uniforms from the returned list
-  _extractAnimatedUniforms(uniforms) {
-    let foundAnimated = false;
-
-    // Keep our animatedUniforms map up-to-date
-    for (const uniformName in uniforms) {
-      const newValue = uniforms[uniformName];
-      if (typeof newValue === 'function') {
-        this.animatedUniforms[uniformName] = newValue;
-        foundAnimated = true;
-      } else {
-        delete this.animatedUniforms[uniformName];
-      }
-    }
-
-    // Update animated flag: `Model` is animated if any uniforms are animated (i.e. functions)
-    this.animated = !isObjectEmpty(this.animatedUniforms);
-
-    if (!foundAnimated) {
-      return uniforms;
-    }
-
-    // If animated uniforms were found, remove them from ordinary uniform list
-    // `Program` class can't (and shouldn't) handle function valued uniforms
-    const staticUniforms = {};
-    for (const uniformName in uniforms) {
-      if (!this.animatedUniforms[uniformName]) {
-        staticUniforms[uniformName] = uniforms[uniformName];
-      }
-    }
-    return staticUniforms;
   }
 
   // Transform Feedback
@@ -428,70 +320,5 @@ count: ${this.stats.profileFrameCount}`
     }
 
     return buffers;
-  }
-
-  _logDrawCallStart(priority) {
-    const logDrawTimeout = priority > 3 ? 0 : LOG_DRAW_TIMEOUT;
-    if (log.priority < priority || Date.now() - this.lastLogTime < logDrawTimeout) {
-      return undefined;
-    }
-
-    this.lastLogTime = Date.now();
-
-    log.group(LOG_DRAW_PRIORITY, `>>> DRAWING MODEL ${this.id}`, {collapsed: log.priority <= 2})();
-
-    return priority;
-  }
-
-  _logDrawCallEnd(priority, vertexArray, uniforms, framebuffer) {
-    // HACK: priority === undefined means logDrawCallStart didn't run
-    if (priority === undefined) {
-      return;
-    }
-
-    const attributeTable = getDebugTableForVertexArray({
-      vertexArray,
-      header: `${this.id} attributes`,
-      attributes: this._attributes
-    });
-
-    const {table: uniformTable, unusedTable, unusedCount} = getDebugTableForUniforms({
-      header: `${this.id} uniforms`,
-      program: this.program,
-      uniforms: Object.assign({}, this.program.uniforms, uniforms)
-    });
-
-    // log missing uniforms
-    const {table: missingTable, count: missingCount} = getDebugTableForUniforms({
-      header: `${this.id} uniforms`,
-      program: this.program,
-      uniforms: Object.assign({}, this.program.uniforms, uniforms),
-      undefinedOnly: true
-    });
-
-    if (missingCount > 0) {
-      log.log('MISSING UNIFORMS', Object.keys(missingTable))();
-      // log.table(priority, missingTable)();
-    }
-    if (unusedCount > 0) {
-      log.log('UNUSED UNIFORMS', Object.keys(unusedTable))();
-      // log.log(priority, 'Unused uniforms ', unusedTable)();
-    }
-
-    const configTable = getDebugTableForProgramConfiguration(this.vertexArray.configuration);
-
-    log.table(priority, attributeTable)();
-
-    log.table(priority, uniformTable)();
-
-    log.table(priority + 1, configTable)();
-
-    logModel(this, uniforms);
-
-    if (framebuffer) {
-      framebuffer.log({priority: LOG_DRAW_PRIORITY, message: `Rendered to ${framebuffer.id}`});
-    }
-
-    log.groupEnd(LOG_DRAW_PRIORITY, `>>> DRAWING MODEL ${this.id}`)();
   }
 }
